refactor(tests): consolidate compileAndRun helpers into tests/helpers/compile.ts (#3109)
Slice 1 of #3109. Extract the three highest-duplication identical-body
compileAndRun clusters into one shared module and migrate 19 test files to
import it under an alias (call sites unchanged). Pure dedup, zero behavior
change — no src/ touched.

- compileAndRunStubs (9): assert success + bare env.console_log_* stub imports
- compileAndRunImportObject (5): assert success + result.importObject!
- compileAndRunBuildImports (5): guard binary + full buildImports host object

Kept as three distinct helpers (not merged) because the clusters link
different host-import sets; merging would drift runtime behavior. Orphaned
compile/buildImports imports dropped where the local helper was their only use.

Parity proof: the 19 files run with vitest --reporter=json before and after
migration yield an identical per-test result set (224 tests, 125 pass / 99
fail; the 99 pre-existing main failures unchanged), at both the original base
and after merging origin/main. tsc --noEmit, biome lint, prettier --check all
green. Net -247 LOC across the 19 files; +78 for the shared module.


✓
