+++ tests/issue-146.test.ts
@@ -1,16 +1,6 @@
 import { describe, it, expect } from "vitest";
 import { compile } from "../src/index.js";
-import { buildImports } from "../src/runtime.js";
-
-async function compileAndRun(source: string) {
-  const result = await compile(source, { fileName: "test.ts" });
-  if (!result.binary || result.binary.length === 0) {
-    throw new Error(`Compile failed:\n${result.errors.map((e) => `  L${e.line}: ${e.message}`).join("\n")}`);
-  }
-  const imports = buildImports(result.imports, undefined, result.stringPool);
-  const { instance } = await WebAssembly.instantiate(result.binary, imports);
-  return instance.exports as Record<string, Function>;
-}
+import { compileAndRunBuildImports as compileAndRun } from "./helpers/compile.js";
 
 async function compileOnly(source: string) {
   const result = await compile(source, { fileName: "test.ts" });
