+++ tests/issue-2065.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { compile } from "../src/index.js";
+import { compileAndRunImportObject as compileAndRun } from "./helpers/compile.js";
 
 // #2065 — the for-of array fast path hoisted the vec's `length` and `data` into
 // locals once before the loop, so elements pushed during iteration were never
@@ -7,15 +7,6 @@ import { compile } from "../src/index.js";
 // array left `data` stale). When the iterable is a plain identifier whose array
 // the body may mutate, the loop now re-reads length/data from the vec each
 // iteration. Non-mutating loops keep the hoisted fast path.
-
-async function compileAndRun(source: string) {
-  const result = await compile(source);
-  expect(result.success, `Compile failed:\n${result.errors.map((e) => `  L${e.line}: ${e.message}`).join("\n")}`).toBe(
-    true,
-  );
-  const { instance } = await WebAssembly.instantiate(result.binary, result.importObject!);
-  return instance.exports as Record<string, Function>;
-}
 
 describe("#2065 for-of observes mid-iteration array mutation", () => {
   it("push during iteration is visited", async () => {
